Added session timeout and automatic redirecting on visiting the site home.
previouly, the only way that the session would be unset would be for the user to
press the logout button, even if the browser window is closed and opened, the session
is not unset. To avoid this, I have implemented a 10 minute session timer.

If the user is inactive for 10 minutes, the session is automatically destroyed.

Also, when the user tries to visit the site home when he/she is loggedin, he will
be redirected to the summary page.