[major]
1. Changed the query for getting the data from all the tables.

The previous returned the same column (companyid) which was being used
to join all the tables again and again. This was causing problems in the name
of the post variables that were being generated.

So changed the query to natural join, which removes the duplicate columns.

2. Added the interface for editing. Not yet added anything else.

The data is being submitted.

3. Changed the submit function that will be called when the submit
is clicked inside the form